backend/requirements.txt
Flask>=2.2
Flask-Cors>=3.0
PyJWT>=2.8
